requirements.txt
python-telegram-bot==11.1.0
pytz==2018.5
requests==2.20.0